Rename State as ViewState class
