+++ app/src/main/java/com/yesferal/hornsapp/app/presentation/MainActivity.kt
@@ -3,7 +3,7 @@ package com.yesferal.hornsapp.app.presentation
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.yesferal.hornsapp.app.R
-import com.yesferal.hornsapp.app.common.transparentStateBar
+import com.yesferal.hornsapp.app.util.transparentStateBar
 import com.yesferal.hornsapp.app.presentation.concert.ConcertsFragment
 
 class MainActivity : AppCompatActivity() {
