+++ app/src/main/java/com/yesferal/hornsapp/app/presentation/common/BasePresenter.kt
@@ -1,15 +1,15 @@
 package com.yesferal.hornsapp.app.presentation.common
 
 @Suppress("UNCHECKED_CAST")
-abstract class BasePresenter<V: BaseContract.View, VD: ViewData>
+abstract class BasePresenter<VIEW: BaseContract.View, VIEW_DATA: ViewData>
     : BaseContract.ActionListener {
-    var view: V? = null
+    var view: VIEW? = null
 
     abstract fun onViewCreated()
-    abstract fun render(state: State<VD>)
+    abstract fun render(state: ViewState<VIEW_DATA>)
 
     override fun attach(view: BaseContract.View) {
-        this.view = view as V
+        this.view = view as VIEW
     }
 
     override fun detachView() {
