+++ app/src/main/java/com/yesferal/hornsapp/app/presentation/concert/ConcertsPresenter.kt
@@ -1,7 +1,7 @@
 package com.yesferal.hornsapp.app.presentation.concert
 
 import com.yesferal.hornsapp.app.presentation.common.BasePresenter
-import com.yesferal.hornsapp.app.presentation.common.State
+import com.yesferal.hornsapp.app.presentation.common.ViewState
 import com.yesferal.hornsapp.domain.usecase.GetConcertsUseCase
 
 class ConcertsPresenter(
@@ -12,7 +12,7 @@ class ConcertsPresenter(
         getConcertsUseCase(
             onSuccess = { list ->
                 val viewData = ConcertViewData(list)
-                val success = State.Success(viewData)
+                val success = ViewState.Success(viewData)
                 render(state = success)
             },
             onError = {
@@ -21,9 +21,9 @@ class ConcertsPresenter(
         )
     }
 
-    override fun render(state: State<ConcertViewData>) {
+    override fun render(state: ViewState<ConcertViewData>) {
         when(state) {
-            is State.Success -> {
+            is ViewState.Success -> {
                 view?.show(concerts = state.viewData.list)
             }
         }
