+++ app/src/main/java/com/yesferal/hornsapp/app/util/Activity.kt
@@ -1,4 +1,4 @@
-package com.yesferal.hornsapp.app.common
+package com.yesferal.hornsapp.app.util
 
 import android.app.Activity
 import android.graphics.Color
